requirement.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Cython==0.29.28
dj-database-url==0.5.0
Django==4.0.3
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
easyocr==1.4.2
gunicorn==20.1.0
idna==3.3
imageio==2.16.2
importlib-metadata==4.11.3
Markdown==3.3.6
networkx==2.8
numpy==1.22.3
opencv-python-headless==4.5.4.60
packaging==21.3
Pillow==9.0.1
psycopg2==2.9.3
PyJWT==2.3.0
pyparsing==3.0.7
python-bidi==0.4.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
requests==2.27.1
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
sqlparse==0.4.2
tesseract==0.1.3
tifffile==2022.4.8
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
zipp==3.7.0
